+++ packages/core/react/src/index.ts
@@ -1,4 +1,5 @@
 'use client'
 
 export * from '@nsui/button'
+export * from '@nsui/spinner'
 export * from '@nsui/provider'