+++ packages/ui/button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nsui/button
 
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @nsui/provider@1.3.0
+
 ## 1.2.0
 
 ### Minor Changes